+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 - [#86](https://github.com/zendframework/zend-expressive-tooling/pull/86) adds the ability to provide the `--modules-path` option to the various
   `module:*` commands via configuration. In each case, if the option is omitted,
-  the command will search for a `Zend\Expressive\Tooling\Module\CommandCommonOptions.modules_path`
+  the command will search for a `Zend\Expressive\Tooling\Module\CommandCommonOptions.--modules-path`
   configuration entry, and use it if present. When present, calling any of these
   commands can now omit the `--modules-path` option when a custom path is
   required.